Move flatmap and map rdds to their own src
